+++ NEWS.md
@@ -1,3 +1,10 @@
-# FPSurveyR (development version)
+# FPSurveyR 1.1.0
+
+* Addition of `fps_data_to_binary` function to compliment overhaul of pipeline's validation and formatting scripts.
+* Update existing functions to account for new "_v" format response columns.
+* Update existing functions to work with new lists in pipeline (e.g. questions - row number lookup in dataset script).
+* Allow running of chart and narrative functions when questions are missing (outputs markdown-formatted message instead).
+
+# FPSurveyR 1.0.0
 
 * Initial release of FPSurveyR to Github.